Add files via upload
In this R script file, I look at what really differentiates a tibble from a normal dataframe. On printing out the output of each object, I realize that a tibble object shows the type of each variable whereas the dataframe object does not. To check out the type/class of each variable in a dataframe, you have to use the str() function.