Add files via upload
This is a notebook that I did when I was learning how to work with data in Python, specifically text files. Today I am going through it to review what I was doing then when I was a newbie. Honestly, there is a lot am (re)learning a lot from it. You should always store your work for future referencing, learning and sharing.